feat: restructure search-wizard directory
