+++ frontend/components/search-wizard/utils.ts
@@ -1,4 +1,30 @@
+import type { IntakeSessionQuestion } from "@services/intake-sessions";
+
 import type { AnswerValue, SummaryRow, WizardAnswers, WizardQuestion } from "./types";
+
+export const mapApiQuestionToWizardQuestion = (
+  question: IntakeSessionQuestion,
+): WizardQuestion => {
+  const normalizedType = question.type.trim().toLowerCase();
+
+  if (normalizedType === "text" || normalizedType === "textarea") {
+    return {
+      id: question.key,
+      kind: "text",
+      title: question.text,
+      description: "",
+      required: true,
+    };
+  }
+
+  return {
+    id: question.key,
+    kind: "text",
+    title: question.text,
+    description: "",
+    required: true,
+  };
+};
 
 export const formatRangeValue = (value: number, unit?: string) => {
   if (unit === "$") {
